Refactor book

+++ LMS.Infrastructure/Repositories/BookRepository.cs
@@ -11,104 +11,51 @@ namespace LMS.Infrastructure.Repositories
         private readonly ApplicationDbContext _context;
         private readonly int _booksToTake = 100;
 
-        public BookRepository(ApplicationDbContext context)
-        {
-            _context = context;
-        }
+        public BookRepository(ApplicationDbContext context) => _context = context;
 
-        public async Task<Book> CreateAsync(Book book, List<int> authors)
+        public async Task<Book> CreateAsync(Book book)
         {
-            using(var transaction = _context.Database.BeginTransaction())
+            try
             {
-                try
-                {
-                    _context.Books.Add(book);
-                    _context.SaveChanges();
+                _context.Books.Add(book);
+                await _context.SaveChangesAsync();
 
-                    #region Authorship
-                    if (book is not null)
-                    {
-                        foreach (int authorId in authors)
-                        {
-                            Authorship authorship = new()
-                            {
-                                BookId = book.Id,
-                                AuthorId = authorId
-                            };
-                            _context.Authorships.Add(authorship);
-                            await _context.SaveChangesAsync();
-                        }
-                    }
-                    #endregion
+                return book;
 
-                    transaction.Commit();
+            } 
+            catch (Exception ex)
+            {
+                Log.Error(ex, ex.Message);
 
-                    return book;
-
-                } catch (Exception ex)
-                {
-                    transaction.Rollback();
-
-                    Log.Error(ex, ex.Message);
-                    await Log.CloseAndFlushAsync();
-
-                    return null;
-                }
+                return null;
             }
         }
 
-        public async Task<Book> UpdateAsync(Book book, List<int> authors)
+        public async Task<Book> UpdateAsync(Book book)
         {
-            using (var transaction = _context.Database.BeginTransaction())
+            try
             {
-                try
-                {
+                Book existingBook = await GetByAsync(book.Id);
+                if (existingBook is null) return new Book();
 
-                    var authorships = await GetAuthorshipByBookIdAsync(book.Id);
+                existingBook.Title = book.Title;
+                existingBook.Description = book.Description;
+                existingBook.Edition = book.Edition;
+                existingBook.ISBN = book.ISBN;
+                existingBook.CategoryId = book.CategoryId;
+                existingBook.ImageUrl = book.ImageUrl;
+                existingBook.Price = book.Price;
+                existingBook.Status = book.Status;
+                existingBook.CopiesAvailable = book.CopiesAvailable;
 
-                    if (authorships is not null)
-                    {
-                        _context.Authorships.RemoveRange(authorships.ToArray());
-                        _context.SaveChanges();
-                        foreach (int authorId in authors)
-                        {
-                            _context.Authorships
-                                .Add(
-                                    new Authorship()
-                                    {
-                                        BookId = book.Id,
-                                        AuthorId = authorId
-                                    });
-                            await _context.SaveChangesAsync();
-                        }
-                    }
-
-                    Book existingBook = await GetByAsync(book.Id);
-                    if(existingBook is not null)
-                    {
-                        existingBook.Title = book.Title;
-                        existingBook.Description = book.Description;
-                        existingBook.Edition = book.Edition;
-                        existingBook.ISBN = book.ISBN;
-                        existingBook.CategoryId = book.CategoryId;
-                        existingBook.ImageUrl = book.ImageUrl;
-                        existingBook.Price = book.Price;
-
-                        _context.Books.Update(existingBook);
-                        await _context.SaveChangesAsync();
-                    }
-
-                    transaction.Commit();
-
-                    return existingBook;
-                }
-                catch (Exception ex)
-                {
-                    transaction.Rollback();
-                    Log.Error(ex, ex.Message);
-                    await Log.CloseAndFlushAsync();
-                    return null;
-                }
+                _context.Entry(existingBook).State = EntityState.Modified;
+                await _context.SaveChangesAsync();
+                return existingBook;
+            }
+            catch (Exception ex)
+            {
+                Log.Error(ex, ex.Message);
+                return new Book();
             }
         }
 
@@ -117,17 +64,16 @@ namespace LMS.Infrastructure.Repositories
             try
             {
                 Book book = await GetByAsync(id);
-
+                if (book is null) return false;
+                
                 _context.Books.Remove(book);
                 await _context.SaveChangesAsync();
-
                 return true;
-
             }
             catch (Exception ex)
             {
                 Log.Error(ex, ex.Message);
-                await Log.CloseAndFlushAsync();
+
                 return false;
             }
             
@@ -138,51 +84,39 @@ namespace LMS.Infrastructure.Repositories
             return await _context.Books
                         .Where(b => b.Id == id)
                         .Include(b => b.Category)
-                        .Include(b => b.Authorships)
-                        .ThenInclude(b => b.Authors)
                         .AsNoTrackingWithIdentityResolution()
                         .FirstOrDefaultAsync();
         }
 
-        private async Task<List<Authorship>> GetAuthorshipByBookIdAsync(int id)
-        {
-            return await _context.Authorships
-                        .Where(auth => auth.BookId == id)
-                        .AsNoTrackingWithIdentityResolution()
-                        .ToListAsync();
-        }
-
-        public async Task<IEnumerable<Book>> GetAllAsync()
+        public async Task<IEnumerable<Book>> GetAsync()
         {
             return await _context.Books
                         .Include(b => b.Category)
-                        .Include(b => b.Authorships)
-                        .ThenInclude(b => b.Authors)
-                        .Include(b => b.PurchaseItems)
-                        .ThenInclude(b => b.Purchase)
                         .Skip(0)
                         .Take(_booksToTake)
                         .OrderBy(b => b.Title)
                         .ToListAsync();
         }
 
-        public async Task<IEnumerable<Book>> GetAllByAsync(string title)
+        public async Task<IEnumerable<Book>> GetAsync(string title)
         {
             return await _context.Books
                         .Where(b => b.Title.Contains(title))
                         .Include(b => b.Category)
                         .Skip(0)
                         .Take(_booksToTake)
+                        .OrderBy(b => b.Title)
                         .ToListAsync();
         }
 
-        public async Task<IEnumerable<Book>> GetAllByAsync(int category)
+        public async Task<IEnumerable<Book>> GetAsync(int category)
         {
             return await _context.Books
                         .Where(b => b.CategoryId == category)
                         .Include(b => b.Category)
                         .Skip(0)
                         .Take(_booksToTake)
+                        .OrderBy(b => b.Title)
                         .ToListAsync();
         }
